requirements.txt
certifi==2019.3.9
chardet==3.0.4
idna==2.8
numpy==1.16.2
Pillow==6.0.0
pip==19.0.3
requests==2.21.0
rockethub==0.1.1
setuptools==41.0.0
six==1.12.0
torch==1.0.1.post2
torchvision==0.2.2.post3
tqdm==4.31.1
urllib3==1.24.1
wheel==0.33.1